create robert lib plus some tidy up
