--- ipxe.cfg
@@ -1,31 +1,0 @@
-#!ipxe
-
-dhcp
-echo "I am ${mac}"
-sleep 9
-chain http://boot.ipxe.org/de
-
-
-set bob http://127.0.0.1/fcos
-set KERNEL ${bob}/fcos-40-kernel
-set INITRD ${bob}/fcos-40-initramfs.img
-set ROOTFS ${bob}/fcos-40-rootfs.img
-set IGNITION ${bob}/bootstrap.ign
-
-kernel ${KERNEL} initrd=main coreos.live.rootfs_url-${ROOTFS} coreos.inst.install_dev=/dev/sda coreos.inst.ignition_url=${IGNITION}
-initrd --name main ${INITRD}
-
-boot
-
-
-##################
-
-set mirror http://mirror.ssvc/alpine
-set bob http://192.168.0.0/alpine
-set alpine ${bob}/v3.20/main
-set flavor virt
-
-kernel ${bob}/vmlinuz-${flavor} console=tty0 modules=loop,squashfs quiet nomodeset modloop=${bob}/modloop-${flavor} alpine_repo=${mirror} initrd=initramfs-${flavor} apkvol=${bob}/headless.apkvol.tar.gz
-initrd ${bob}/initramfs-${flavor}
-
-boot
