--- opt_bob/wendy/wsgi.py
@@ -1,4 +1,0 @@
-from wendy import app
-
-if __name__ == "__main__":
-    app.run()
